+++ internal/ai/train/annealer.go
@@ -3,14 +3,14 @@ package train
 import "math"
 
 type Annealer struct {
-	StartTemp float64
-	MinTemp   float64
-	Decay     float64
+	StartTemp float32
+	MinTemp   float32
+	Decay     float32
 	WarmUp    int // Number of epochs to stay at StartTemp
 }
 
 // GetTemp calculates the temperature for the current epoch
-func (a *Annealer) GetTemp(epoch int) float64 {
+func (a *Annealer) GetTemp(epoch int) float32 {
 	// Phase 1: Warm-up (Hold temperature constant)
 	if epoch < a.WarmUp {
 		return a.StartTemp
@@ -19,7 +19,7 @@ func (a *Annealer) GetTemp(epoch int) float64 {
 	// Phase 2: Exponential Decay
 	// Subtract warm-up epochs so decay starts from the peak
 	decayEpoch := float64(epoch - a.WarmUp)
-	temp := a.StartTemp * math.Pow(a.Decay, decayEpoch)
+	temp := float32(float64(a.StartTemp) * math.Pow(float64(a.Decay), decayEpoch))
 
 	if temp < a.MinTemp {
 		return a.MinTemp
